Add multithreaded compression.
This adds a StreamOptimizedCopyDisk() function to sparse.c which will copy a whole disk.
A number of threads is created, which run in deflateGrainThread(), using a GrainThreadContext
to keep track of the current input and output positions. Each thread reads a grain from the
input stream at the current input position, after advancing the position so another thread
can start reading at the updated position. The grain is then compressed, which is CPU intensive.
When done, the output position is advanced, before writing the compressed grain.

The number of threads is configurable (-n option), and defaults to the number of available CPUs.
The number can also be set with the environment variable VMDKCONVERT_NUM_THREADS.